Build ntdll.lib import library
Wine's dbghelp uses a few functions from ntdll.dll - however, an
import library for it is available in Windows SDKs only starting with
the SDK for Windows 8.0. Rather than force using certain versions of
the Windows SDK, build a .lib file for just the 5 functions that
Wine's dbghelp uses.

For simplicity, we use one .def file for both 32-bit and 64-bit
imports, to avoid having a separate .def file for each target
platform.